demo-api/app/requirements.txt
# Core backend
django==4.1.1
djangorestframework==3.14.0
django-filter==22.1

# Webserver
gunicorn==20.1.0